+++ app/src/main/java/com/jsadevtech/jsa/vgc/auxiliars/EventoBD.java
@@ -1,4 +1,4 @@
-package com.jsadevtech.jsa.vgc;
+package com.jsadevtech.jsa.vgc.auxiliars;
 
 import com.jsadevtech.jsa.vgc.exceptions.ConnectionFailedException;
 import com.jsadevtech.jsa.vgc.exceptions.CouldNotGetInformationException;
