+++ app/src/main/java/com/jsadevtech/jsa/vgc/auxiliars/Group.java
@@ -1,4 +1,4 @@
-package com.jsadevtech.jsa.vgc;
+package com.jsadevtech.jsa.vgc.auxiliars;
 
 import java.util.ArrayList;
 import java.util.List;
