+++ app/src/main/java/com/jsadevtech/jsa/vgc/auxiliars/InvitadoBD.java
@@ -1,4 +1,4 @@
-package com.jsadevtech.jsa.vgc;
+package com.jsadevtech.jsa.vgc.auxiliars;
 
 import com.jsadevtech.jsa.vgc.exceptions.ConnectionFailedException;
 import com.jsadevtech.jsa.vgc.exceptions.CouldNotConvertFormatException;
