+++ app/src/main/java/com/jsadevtech/jsa/vgc/screens/comprar_online_screen.java
@@ -1,4 +1,4 @@
-package com.jsadevtech.jsa.vgc;
+package com.jsadevtech.jsa.vgc.screens;
 
 import android.app.Activity;
 import android.os.Bundle;
@@ -8,6 +8,8 @@ import android.view.MenuItem;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
+
+import com.jsadevtech.jsa.vgc.R;
 
 /**
  * Created by Ivan on 10/07/2015.
