+++ app/src/main/java/com/jsadevtech/jsa/vgc/screens/horarios_screen.java
@@ -1,10 +1,12 @@
-package com.jsadevtech.jsa.vgc;
+package com.jsadevtech.jsa.vgc.screens;
 
 import android.app.Activity;
 import android.os.Bundle;
 import android.os.StrictMode;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import com.jsadevtech.jsa.vgc.R;
 
 /**
  * Created by Ivan on 01/07/2015.
