+++ app/src/main/java/com/jsadevtech/jsa/vgc/screens/plano_screen.java
@@ -1,31 +1,34 @@
-package com.jsadevtech.jsa.vgc;
+package com.jsadevtech.jsa.vgc.screens;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.util.SparseArray;
+import android.os.StrictMode;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.widget.ExpandableListView;
 
+import com.jsadevtech.jsa.vgc.R;
 
-public class informacion extends Activity {
-    // more efficient than HashMap for mapping integers to objects
-    private SparseArray<Group> groups = new SparseArray<>();
+/**
+ * Created by Ivan on 01/07/2015.
+ */
+public class plano_screen extends Activity {
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState){
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_informacion);
-        //StrictMode.enableDefaults();//modo stricto necesario para la conexion a internet
+        setContentView(R.layout.activity_plano_screen);
+        StrictMode.enableDefaults();//modo stricto necesario para la conexion a internet
+
     }
 
 
-        @Override
+    @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_informacion, menu);
+        getMenuInflater().inflate(R.menu.menu_main_screen, menu);
         return true;
     }
+
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
@@ -41,4 +44,5 @@ public class informacion extends Activity {
 
         return super.onOptionsItemSelected(item);
     }
+
 }
