+++ app/src/main/res/layout/activity_comprar_fisico_screen.xml
@@ -7,7 +7,7 @@
     <fragment xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
         android:layout_height="@dimen/fragment_map_height" android:id="@+id/map"
-        tools:context="com.jsadevtech.jsa.vgc.comprar_fisico_screen"
+        tools:context="com.jsadevtech.jsa.vgc.screens.comprar_fisico_screen"
         android:name="com.google.android.gms.maps.SupportMapFragment" />
 
     <RelativeLayout
